+++ composeApp/build.gradle.kts
@@ -46,8 +46,8 @@ kotlin {
             implementation(projects.kPermissionsCamera)
             implementation(projects.kPermissionsStorage)
             implementation(projects.kPermissionsGallery)
-            implementation(projects.kPermissionsAudio)
-            implementation(projects.kPermissionsVideo)
+            implementation(projects.kPermissionsReadAudio)
+            implementation(projects.kPermissionsReadVideo)
             implementation(projects.kPermissionCMPLocationWhenInUse)
             implementation(projects.kpermissionsCMPLocationAlways)
             implementation(projects.kpermissionsLocationChecker)
