--- kPermissionsAudio/src/commonMain/kotlin/io/github/kPermissionsAudio/ReadAudioPermission.kt
@@ -1,5 +1,0 @@
-package io.github.kPermissionsAudio
-
-import io.github.kPermissions_api.Permission
-
-expect object ReadAudioPermission : Permission
