+++ kPermissionsCamera/src/androidMain/kotlin/io/github/kpermissionsCamera/CameraPermission.android.kt
@@ -6,7 +6,7 @@ import io.github.kPermissions_api.Permission
 
 
 actual object CameraPermission : Permission {
-    override val name: String
+    actual override val name: String
         get() = "camera"
 
     @Suppress("RedundantNullableReturnType")
@@ -16,18 +16,18 @@ actual object CameraPermission : Permission {
     private var _minSdk: Int? = null
     private var _maxSdk: Int? = null
 
-    override val minSdk: Int?
+    actual override val minSdk: Int?
         get() = _minSdk
 
-    override val maxSdk: Int?
+    actual override val maxSdk: Int?
         get() = _maxSdk
 
-    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
         _minSdk = minSdk
         _maxSdk = maxSdk
     }
 
-    override suspend fun isServiceAvailable(): Boolean {
+    actual override suspend fun isServiceAvailable(): Boolean {
 
         return AppContextProvider.appContext.packageManager.hasSystemFeature(PackageManager.FEATURE_CAMERA_ANY)
     }
