+++ kPermissionsCamera/src/commonMain/kotlin/io/github/kpermissionsCamera/CameraPermission.kt
@@ -3,4 +3,10 @@ package io.github.kpermissionsCamera
 import io.github.kPermissions_api.Permission
 
 
-expect object CameraPermission : Permission
+expect object CameraPermission : Permission {
+    override val name: String
+    override val minSdk: Int?
+    override val maxSdk: Int?
+    override suspend fun isServiceAvailable(): Boolean
+    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?)
+}
