+++ kPermissionsCamera/src/iosMain/kotlin/io/github/kpermissionsCamera/CameraPermission.ios.kt
@@ -35,12 +35,12 @@ internal fun cameraPermissionRequest(): ((Boolean) -> Unit) -> Unit = { callback
 }
 
 actual object CameraPermission : Permission {
-    override val name: String
+    actual override val name: String
         get() = "camera"
     override val permissionRequest: ((Boolean) -> Unit) -> Unit
         get() = cameraPermissionRequest()
 
-    override suspend fun isServiceAvailable(): Boolean {
+    actual override suspend fun isServiceAvailable(): Boolean {
         return AVCaptureDevice.defaultDeviceWithMediaType(AVMediaTypeVideo) != null
     }
 
@@ -51,14 +51,14 @@ actual object CameraPermission : Permission {
     private var _minSdk: Int? = null
     private var _maxSdk: Int? = null
 
-    override val minSdk: Int?
+    actual override val minSdk: Int?
         get() = _minSdk
 
-    override val maxSdk: Int?
+    actual override val maxSdk: Int?
         get() = _maxSdk
 
 
-    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
         _minSdk = minSdk
         _maxSdk = maxSdk
     }
