+++ kPermissionsGallery/src/androidMain/kotlin/io/github/kPermissionsGallery/GalleryPermission.android.kt
@@ -6,7 +6,7 @@ import io.github.kPermissions_api.Permission
 
 
 actual object GalleryPermission : Permission {
-   override val name: String
+    actual override val name: String
       get() = "gallery"
    override val androidPermissionName: String?
       get() = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) Manifest.permission.READ_MEDIA_IMAGES else null
@@ -14,19 +14,19 @@ actual object GalleryPermission : Permission {
    private var _minSdk: Int? = null
    private var _maxSdk: Int? = null
 
-   override val minSdk: Int?
+    actual override val minSdk: Int?
       get() = _minSdk
 
-   override val maxSdk: Int?
+    actual override val maxSdk: Int?
       get() = _maxSdk
 
 
-   override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
       _minSdk = minSdk
       _maxSdk = maxSdk
    }
 
-   override suspend fun isServiceAvailable(): Boolean {
+    actual override suspend fun isServiceAvailable(): Boolean {
       return true
    }
