+++ kPermissionsNotification/src/commonMain/kotlin/io/github/kpermissionnotification/NotificationPermission.kt
@@ -2,4 +2,10 @@ package io.github.kpermissionnotification
 
 import io.github.kPermissions_api.Permission
 
-expect object NotificationPermission : Permission
+expect object NotificationPermission : Permission {
+    override val name: String
+    override val minSdk: Int?
+    override val maxSdk: Int?
+    override suspend fun isServiceAvailable(): Boolean
+    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?)
+}
