+++ kPermissionsNotification/src/iosMain/kotlin/io/github/kpermissionnotification/NotificationPermission.ios.kt
@@ -17,19 +17,19 @@ import platform.darwin.dispatch_semaphore_wait
 
 actual object NotificationPermission : Permission {
 
-    override val name: String = "notification"
+    actual override val name: String = "notification"
 
     private var _minSdk: Int? = null
     private var _maxSdk: Int? = null
-    override val minSdk: Int? get() = _minSdk
-    override val maxSdk: Int? get() = _maxSdk
+    actual override val minSdk: Int? get() = _minSdk
+    actual override val maxSdk: Int? get() = _maxSdk
 
-    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
         _minSdk = minSdk
         _maxSdk = maxSdk
     }
 
-    override suspend fun isServiceAvailable(): Boolean {
+    actual override suspend fun isServiceAvailable(): Boolean {
         return true
     }
