+++ kPermissionsReadAudio/src/androidMain/AndroidManifest.xml
