+++ kPermissionsReadAudio/src/commonMain/kotlin/io/github/kPermissionsAudio/ReadAudioPermission.kt
@@ -1,0 +1,11 @@
+package io.github.kPermissionsAudio
+
+import io.github.kPermissions_api.Permission
+
+expect object ReadAudioPermission : Permission {
+    override val name: String
+    override val minSdk: Int?
+    override val maxSdk: Int?
+    override suspend fun isServiceAvailable(): Boolean
+    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?)
+}
