+++ kPermissionsReadVideo/.gitignore
