+++ kPermissionsReadVideo/src/androidMain/AndroidManifest.xml
