+++ kPermissionsReadVideo/src/androidMain/kotlin/io/github/kPermissionsVideo/ReadVideoPermission.android.kt
@@ -1,0 +1,40 @@
+package io.github.kPermissionsVideo
+
+import android.Manifest
+import android.os.Build
+import io.github.kPermissions_api.Permission
+import io.github.kpermissions_cmp.PlatformIgnore
+import io.github.kpermissions_cmp.setIgnore
+
+actual object ReadVideoPermission : Permission {
+    init {
+        this.setIgnore(PlatformIgnore.IOS)
+    }
+
+    actual override val name: String
+        get() = "read_video"
+    override val androidPermissionName: String?
+        get() = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) Manifest.permission.READ_MEDIA_VIDEO else null
+
+    private var _minSdk: Int? = null
+    private var _maxSdk: Int? = null
+
+    actual override val minSdk: Int?
+        get() = _minSdk
+
+    actual override val maxSdk: Int?
+        get() = _maxSdk
+
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+        _minSdk = minSdk
+        _maxSdk = maxSdk
+    }
+
+    actual override suspend fun isServiceAvailable(): Boolean {
+        return true
+    }
+
+
+
+
+}
