+++ kPermissionsReadVideo/src/commonMain/kotlin/io/github/kPermissionsVideo/ReadVideoPermission.kt
@@ -1,0 +1,11 @@
+package io.github.kPermissionsVideo
+
+import io.github.kPermissions_api.Permission
+
+expect object ReadVideoPermission : Permission {
+    override val name: String
+    override val minSdk: Int?
+    override val maxSdk: Int?
+    override suspend fun isServiceAvailable(): Boolean
+    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?)
+}
