+++ kPermissionsStorage/src/iosMain/kotlin/io/github/kPermissionsStorage/StoragePermission.ios.kt
@@ -10,12 +10,12 @@ actual object WriteStoragePermission : Permission {
         this.setIgnore(PlatformIgnore.IOS)
     }
 
-    override val name: String
+    actual override val name: String
         get() = "write_storage"
     override val permissionRequest: ((Boolean) -> Unit) -> Unit
         get() = {}
 
-    override suspend fun isServiceAvailable(): Boolean {
+    actual override suspend fun isServiceAvailable(): Boolean {
         return true
     }
 
@@ -26,14 +26,14 @@ actual object WriteStoragePermission : Permission {
     private var _minSdk: Int? = null
     private var _maxSdk: Int? = null
 
-    override val minSdk: Int?
+    actual override val minSdk: Int?
         get() = _minSdk
 
-    override val maxSdk: Int?
+    actual override val maxSdk: Int?
         get() = _maxSdk
 
 
-    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
         _minSdk = minSdk
         _maxSdk = maxSdk
     }
@@ -44,12 +44,12 @@ actual object WriteStoragePermission : Permission {
 }
 
 actual object ReadStoragePermission : Permission {
-    override val name: String
+    actual override val name: String
         get() = "read_storage"
     override val permissionRequest: ((Boolean) -> Unit) -> Unit
         get() = {}
 
-    override suspend fun isServiceAvailable(): Boolean {
+    actual override suspend fun isServiceAvailable(): Boolean {
         return true
     }
 
@@ -60,14 +60,14 @@ actual object ReadStoragePermission : Permission {
     private var _minSdk: Int? = null
     private var _maxSdk: Int? = null
 
-    override val minSdk: Int?
+    actual override val minSdk: Int?
         get() = _minSdk
 
-    override val maxSdk: Int?
+    actual override val maxSdk: Int?
         get() = _maxSdk
 
 
-    override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
+    actual override fun setMainAndMaxSdk(minSdk: Int?, maxSdk: Int?) {
         _minSdk = minSdk
         _maxSdk = maxSdk
     }
