--- kPermissionsVideo/build.gradle.kts
@@ -1,160 +1,0 @@
-import com.vanniktech.maven.publish.SonatypeHost
-import org.gradle.nativeplatform.platform.internal.DefaultNativePlatform.getCurrentOperatingSystem
-
-plugins {
-    alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidKotlinMultiplatformLibrary)
-
-//    alias(libs.plugins.composeMultiplatform)
-//    alias(libs.plugins.composeCompiler)
-
-    id("maven-publish")
-    id("signing")
-    alias(libs.plugins.maven.publish)
-}
-
-mavenPublishing {
-    coordinates("io.github.the-best-is-best", "kpermissions-video", "1.0.0")
-
-    publishToMavenCentral(SonatypeHost.S01, true)
-    signAllPublications()
-
-    pom {
-        name.set("KPermissionsVideo")
-        description.set("KPermissionsVideo is a Kotlin Multiplatform module that handles permission for reading videos from media storage. It's part of the [kPermissions](https://github.com/the-best-is-best/kPermissions) ecosystem.")
-        url.set("https://github.com/the-best-is-best/kPermissions")
-        licenses {
-            license {
-                name.set("Apache-2.0")
-                url.set("https://opensource.org/licenses/Apache-2.0")
-            }
-        }
-        issueManagement {
-            system.set("Github")
-            url.set("https://github.com/the-best-is-best/kPermissions/issues")
-        }
-        scm {
-            connection.set("https://github.com/the-best-is-best/kPermissions.git")
-            url.set("https://github.com/the-best-is-best/kPermissions")
-        }
-        developers {
-            developer {
-                id.set("MichelleRaouf")
-                name.set("Michelle Raouf")
-                email.set("eng.michelle.raouf@gmail.com")
-            }
-        }
-    }
-
-}
-
-signing {
-    useGpgCmd()
-    sign(publishing.publications)
-}
-
-tasks.withType<PublishToMavenRepository> {
-    val isMac = getCurrentOperatingSystem().isMacOsX
-    onlyIf {
-        isMac.also {
-            if (!isMac) logger.error(
-                """
-                    Publishing the library requires macOS to be able to generate iOS artifacts.
-                    Run the task on a mac or use the project GitHub workflows for publication and release.
-                """
-            )
-        }
-    }
-}
-
-kotlin {
-
-    // Target declarations - add or remove as needed below. These define
-    // which platforms this KMP module supports.
-    // See: https://kotlinlang.org/docs/multiplatform-discover-project.html#targets
-    androidLibrary {
-        namespace = "io.github.kPermissionsVideo"
-        compileSdk = libs.versions.android.compileSdk.get().toInt()
-        minSdk = libs.versions.android.minSdk.get().toInt()
-
-
-        withHostTestBuilder {
-        }
-
-        withDeviceTestBuilder {
-            sourceSetTreeName = "test"
-        }.configure {
-            instrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        }
-    }
-
-    // For iOS targets, this is also where you should
-    // configure native binary output. For more information, see:
-    // https://kotlinlang.org/docs/multiplatform-build-native-binaries.html#build-xcframeworks
-
-    // A step-by-step guide on how to include this library in an XCode
-    // project can be found here:
-    // https://developer.android.com/kotlin/multiplatform/migrate
-    val xcfName = "kPermissionsVideoKit"
-
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64(),
-    ).forEach { target ->
-        target.binaries.framework {
-            baseName = xcfName
-        }
-    }
-
-    // Source set declarations.
-    // Declaring a target automatically creates a source set with the same name. By default, the
-    // Kotlin Gradle Plugin creates additional source sets that depend on each other, since it is
-    // common to share sources between related targets.
-    // See: https://kotlinlang.org/docs/multiplatform-hierarchy.html
-    sourceSets {
-        commonMain {
-            dependencies {
-                implementation(libs.kotlin.stdlib)
-                // Add KMP dependencies here
-                //   implementation(compose.runtime)
-                api(projects.kPermissionsApi)
-                api(projects.kPermissionsCore)
-                implementation(projects.kPermissionsCMP)
-            }
-        }
-
-        commonTest {
-            dependencies {
-                implementation(libs.kotlin.test)
-            }
-        }
-
-        androidMain {
-            dependencies {
-                // Add Android-specific dependencies here. Note that this source set depends on
-                // commonMain by default and will correctly pull the Android artifacts of any KMP
-                // dependencies declared in commonMain.
-            }
-        }
-
-        getByName("androidDeviceTest") {
-            dependencies {
-                implementation(libs.androidx.runner)
-                implementation(libs.androidx.core)
-                implementation(libs.androidx.testExt.junit)
-            }
-        }
-
-        iosMain {
-            dependencies {
-                // Add iOS-specific dependencies here. This a source set created by Kotlin Gradle
-                // Plugin (KGP) that each specific iOS target (e.g., iosX64) depends on as
-                // part of KMP’s default source set hierarchy. Note that this source set depends
-                // on common by default and will correctly pull the iOS artifacts of any
-                // KMP dependencies declared in commonMain.
-            }
-        }
-    }
-
-}
