--- kPermissionsVideo/src/commonMain/kotlin/io/github/kPermissionsVideo/ReadVideoPermission.kt
@@ -1,5 +1,0 @@
-package io.github.kPermissionsVideo
-
-import io.github.kPermissions_api.Permission
-
-expect object ReadVideoPermission : Permission
