+++ kpermissionsCMPLocationAlways/build.gradle.kts
@@ -10,7 +10,7 @@ plugins {
 }
 
 mavenPublishing {
-    coordinates("io.github.the-best-is-best", "kpermissions-cmp-location-always", "1.0.0")
+    coordinates("io.github.the-best-is-best", "kpermissions-cmp-location-always", "1.0.1")
 
     publishToMavenCentral(SonatypeHost.S01, true)
     signAllPublications()
