--- kpermissionsLocationWhenInUse/src/androidMain/kotlin/io/github/kpermissionslocationWhenInUse/BluetoothStateReceiver.kt
@@ -1,2 +1,0 @@
-package io.github.kpermissionslocationWhenInUse
-
