+++ settings.gradle.kts
@@ -33,8 +33,8 @@ include(":kPermissionsCore")
 include(":kPermissionsCamera")
 include(":kPermissionsStorage")
 include(":kPermissionsGallery")
-include(":kPermissionsAudio")
-include(":kPermissionsVideo")
+include(":kPermissionsReadAudio")
+include(":kPermissionsReadVideo")
 include(":kPermissionsCMP")
 include(":kPermissionsApi")
 include(":kpermissionsLocationChecker")
